sibar them lich su xuat

+++ warehouseqr/src/main/java/com/tttn/warehouseqr/modules/outbound/repository/OutboundReceiptRepository.java
@@ -3,5 +3,5 @@ package com.tttn.warehouseqr.modules.outbound.repository;
 import com.tttn.warehouseqr.modules.outbound.entity.OutboundReceipt;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OutboundReceiptRepository extends JpaRepository<OutboundReceipt, Integer> {
+public interface OutboundReceiptRepository extends JpaRepository<OutboundReceipt, Long> {
 }
